wip: all bindings split
