+++ skymp5-server/cpp/addon/property_bindings/AppearanceBinding.cpp
@@ -1,0 +1,28 @@
+#include "AppearanceBinding.h"
+
+Napi::Value AppearanceBinding::Get(Napi::Env env, ScampServer &scampServer, uint32_t formId) {
+  auto &partOne = scampServer.GetPartOne();
+
+  auto& actor = partOne->worldState.GetFormAt<MpActor>(formId);
+  auto& appearanceDump = actor.GetAppearanceAsJson();
+  if (!appearanceDump.empty()) {
+    return NapiHelper::ParseJson(env, appearanceDump);
+  } else {
+    return env.Null();
+  }
+}
+
+void AppearanceBinding::Set(Napi::Env env, ScampServer &scampServer, uint32_t formId, Napi::Value newValue) {
+  // TODO: Live update of appearance
+  // TODO: Validation
+  auto &partOne = scampServer.GetPartOne();
+  auto& actor = partOne->worldState.GetFormAt<MpActor>(formId);
+  if (newValue.IsObject()) {
+    auto appearanceDump = NapiHelper::Stringify(env, newValue);
+    simdjson::dom::parser p;
+    auto appearance = Appearance::FromJson(p.parse(appearanceDump));
+    actor.SetAppearance(&appearance);
+  } else {
+    actor.SetAppearance(nullptr);
+  }
+}
