+++ skymp5-server/cpp/addon/property_bindings/EquipmentBinding.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "PropertyBinding.h"
+
+class EquipmentBinding : public PropertyBinding {
+public:
+  std::string GetPropertyName() const override { return "equipment"; }
+  Napi::Value Get(Napi::Env env, ScampServer &scampServer, uint32_t formId) override;
+  void Set(Napi::Env env, ScampServer &scampServer, uint32_t formId, Napi::Value newValue) override;
+};
