+++ skymp5-server/cpp/addon/property_bindings/NeighborsBinding.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "PropertyBinding.h"
+
+class NeighborsBinding : public PropertyBinding {
+public:
+  std::string GetPropertyName() const override { return "neighbors"; }
+  Napi::Value Get(Napi::Env env, ScampServer &scampServer, uint32_t formId) override;
+};
