+++ skymp5-server/cpp/addon/property_bindings/PropertyBinding.h
@@ -1,0 +1,12 @@
+#include <cstdint>
+#include <napi.h>
+#include <string>
+#include "ScampServer.h"
+
+class PropertyBinding {
+public:
+  virtual ~PropertyBinding() = default;
+  virtual std::string GetPropertyName() const = 0;
+  virtual Napi::Value Get(Napi::Env env, ScampServer &scampServer, uint32_t formId);
+  virtual void Set(Napi::Env env, ScampServer &scampServer, uint32_t formId, Napi::Value newValue);
+};
