+++ skymp5-server/cpp/addon/property_bindings/TypeBinding.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "PropertyBinding.h"
+
+class TypeBinding : public PropertyBinding {
+  public:
+  std::string GetPropertyName() const override { return "type"; }
+  Napi::Value Get(Napi::Env env, ScampServer &scampServer, uint32_t formId) override;
+};
